DEV-1216 move close network btn to governance tab (#789)
