+++ components/profile/my-network-settings/logo-and-colours-settings.tsx
@@ -5,7 +5,6 @@ import { useTranslation } from "next-i18next";
 import getConfig from "next/config";
 
 import InfoIconEmpty from "assets/icons/info-icon-empty";
-import LockedIcon from "assets/icons/locked-icon";
 
 import Button from "components/button";
 import AmountCard from "components/custom-network/amount-card";
@@ -14,7 +13,7 @@ import ImageUploader from "components/image-uploader";
 
 import { useAppState } from "contexts/app-state";
 import { useNetworkSettings } from "contexts/network-settings";
-import { addToast, toastError, toastSuccess } from "contexts/reducers/change-toaster";
+import { addToast } from "contexts/reducers/change-toaster";
 
 import { formatDate } from "helpers/formatDate";
 import { getQueryableText, urlWithoutProtocol } from "helpers/string";
@@ -23,7 +22,6 @@ import { MetamaskErrors } from "interfaces/enums/Errors";
 import { Network } from "interfaces/network";
 
 import useApi from "x-hooks/use-api";
-import { useAuthentication } from "x-hooks/use-authentication";
 import useBepro from "x-hooks/use-bepro";
 import { useNetwork } from "x-hooks/use-network";
 
@@ -44,13 +42,11 @@ export default function LogoAndColoursSettings({
 }: Props) {
   const { t } = useTranslation(["common", "custom-network"]);
   const [isRegistering, setIsRegistering] = useState(false);
-  const [isClosing, setIsClosing] = useState(false);
 
   const {state, dispatch} = useAppState();
 
-  const { updateNetwork, processEvent } = useApi();
-  const {  handleAddNetworkToRegistry, handleCloseNetwork } = useBepro();
-  const { updateWalletBalance } = useAuthentication();
+  const { processEvent } = useApi();
+  const { handleAddNetworkToRegistry } = useBepro();
   const { updateActiveNetwork } = useNetwork();
 
   const {
@@ -58,7 +54,6 @@ export default function LogoAndColoursSettings({
     fields,
     settings,
     forcedNetwork,
-    isAbleToClosed,
   } = useNetworkSettings();
 
   const handleColorChange = (value) => fields.colors.setter(value);
@@ -129,46 +124,6 @@ export default function LogoAndColoursSettings({
                       error);
       })
       .finally(() => setIsRegistering(false));
-  }
-
-  function handleCloseMyNetwork() {
-    if (
-      !state.Service?.network?.active ||
-      !state.currentUser?.login ||
-      !state.currentUser?.accessToken ||
-      !state.currentUser?.walletAddress ||
-      !state.Service?.active
-    )
-      return;
-
-    setIsClosing(true);
-
-    handleCloseNetwork()
-      .then(() => {
-        return updateNetwork({
-          githubLogin: state.currentUser.login,
-          isClosed: true,
-          creator: state.currentUser.walletAddress,
-          networkAddress: network?.networkAddress,
-          accessToken: state.currentUser?.accessToken,
-        });
-      })
-      .then(() => {
-        updateWalletBalance(true);
-
-        if (isCurrentNetwork) updateActiveNetwork(true);
-
-        return updateEditingNetwork();
-      })
-      .then(() =>
-        dispatch(toastSuccess(t("custom-network:messages.network-closed"),
-                              t("actions.success"))))
-      .catch((error) =>
-        dispatch(toastError(t("custom-network:errors.failed-to-close-network", { error }),
-                            t("actions.failed"))))
-      .finally(() => {
-        setIsClosing(false);
-      });
   }
 
   return (
@@ -268,25 +223,6 @@ export default function LogoAndColoursSettings({
           `}
             lg
           />
-        </Col>
-
-        <Col xs="auto">
-          <Button
-            color="dark-gray"
-            disabled={!isAbleToClosed || isClosing || !state.currentUser?.login}
-            className="ml-2"
-            onClick={handleCloseMyNetwork}
-          >
-            {(!isAbleToClosed || !state.currentUser?.login) && (
-              <LockedIcon className="me-2" />
-            )}
-            <span>{t("custom-network:close-network")}</span>
-            {isClosing ? (
-              <span className="spinner-border spinner-border-xs ml-1" />
-            ) : (
-              ""
-            )}
-          </Button>
         </Col>
       </Row>
       <Row className="mb-2 justify-content-center">
